+++ packages/story-editor/src/components/style/icons/animationIcon.js
