+++ packages/story-editor/src/components/style/icons/imageSelectionIcon.js
