+++ packages/story-editor/src/components/style/icons/linkIcon.js
