+++ packages/story-editor/src/components/style/icons/multiSelectionIcon.js
