+++ packages/story-editor/src/components/style/icons/selectionIcon.js
