+++ packages/story-editor/src/components/style/icons/shapeSelectionIcon.js
