+++ packages/story-editor/src/components/style/icons/textSelectionIcon.js
