+++ packages/story-editor/src/components/style/icons/videoSelectionIcon.js
